Migrated affinity sprites to separate scriptable object
