+++ Assets/ScriptableObjects/AffinitySprites.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1b49b363b441cee4cbe46d20cd6ced87
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
